re-org'd datalake.py to move job and crawler out to a different module
